requirements.txt
# --- Core runtime ---
python-dotenv==1.0.1

# --- Web / API ---
fastapi==0.128.0
uvicorn==0.40.0
starlette==0.50.0
anyio==4.12.1
h11==0.16.0
sniffio==1.3.1

# --- HTTP / Feeds ---
requests==2.32.3
httpx==0.28.1
httpcore==1.0.9
feedparser==6.0.11
certifi==2026.1.4
idna==3.11

# --- Data / Utils ---
pydantic==2.12.5
pydantic_core==2.41.5
typing_extensions==4.15.0
typing-inspection==0.4.2
tqdm==4.67.1
click==8.3.1
distro==1.9.0

# --- Math / Embeddings (local & future-proof) ---
numpy==2.1.3

# --- OpenAI (future activation, currently optional) ---
openai==2.15.0
jiter==0.12.0
annotated-types==0.7.0